refactor(f2z-kt-core): move the §4.6 authority policy where a client can link it
`KT.md` §8.1 step 7 makes verifying `SignedAuthorityPolicy` a **client**
obligation — "fetch §4.6's `SignedAuthorityPolicy` for this `log_id` and verify
it under the log key already accepted in step 2" — but the structure lived in
`f2z-kt`, which is AGPL-3.0. A client linking it to decode the document would be
relicensed by doing so, and the only other way to discharge the step is a second
decoder for a signed structure, which §11.4's *one crate, three consumers* rule
exists to forbid.

So the structure moves to `f2z-kt-core::policy` and `f2z-kt::policy`
re-exports it. `sign_policy` stays behind: it reads the log's live
`AuthorityConfig` and signs with the log's `LogSigner`, neither of which belongs
on a client.

The strict-verification registry entry moves with the function it names.
